+++ javaScript/projects/testing-practice/src/index.js
@@ -1,5 +1,1 @@
-import test from './components/test';
 import './index.css';
-
-test.print();
-document.body.appendChild(test.render());