+++ javaScript/projects/testing-practice/src/testing/sum.test.js
@@ -1,0 +1,5 @@
+import sum from '../components/sum';
+
+test('Adds 1 + 2 equals 3', () => {
+  expect(sum(1, 2).toBe(3));
+});