+++ javaScript/projects/testing-practice/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 ---
+### 13 FEB 2024
+- Initial commit for `testing-practice`.
+- Installed `babel-jest`/`babel/preset-env`/`babel/core`/`jest`.
+---
 ### 14 JAN 2024
 - Reformatted `CHANGELOG.md`.
 - Updated instructions in `README.md`.